+++ etna/ensembles/voting_ensemble.py
@@ -1,4 +1,6 @@
 from copy import deepcopy
+from typing import Any
+from typing import Dict
 from typing import List
 from typing import Optional
 
@@ -47,7 +49,13 @@ class VotingEnsemble(Pipeline):
 
     support_confidence_interval = False
 
-    def __init__(self, pipelines: List[Pipeline], weights: Optional[List[float]] = None, n_jobs: int = 1):
+    def __init__(
+        self,
+        pipelines: List[Pipeline],
+        weights: Optional[List[float]] = None,
+        n_jobs: int = 1,
+        joblib_params: Dict[str, Any] = dict(verbose=11, backend="multiprocessing", mmap_mode="c"),
+    ):
         """Init VotingEnsemble.
 
         Parameters
@@ -58,6 +66,8 @@ class VotingEnsemble(Pipeline):
             list of pipelines' weights; weights will be normalized automatically.
         n_jobs:
             number of jobs to run in parallel
+        joblib_params:
+            additional parameters for joblib.Parallel
 
         Raises
         ------
@@ -69,6 +79,7 @@ class VotingEnsemble(Pipeline):
         self.weights = self._process_weights(weights=weights, pipelines_number=len(pipelines))
         self.pipelines = pipelines
         self.n_jobs = n_jobs
+        self.joblib_params = joblib_params
 
     @staticmethod
     def _validate_pipeline_number(pipelines: List[Pipeline]):
@@ -116,7 +127,7 @@ class VotingEnsemble(Pipeline):
         VotingEnsemble:
             fitted ensemble
         """
-        self.pipelines = Parallel(n_jobs=self.n_jobs, backend="multiprocessing", verbose=11)(
+        self.pipelines = Parallel(n_jobs=self.n_jobs, **self.joblib_params)(
             delayed(self._fit_pipeline)(pipeline=pipeline, ts=deepcopy(ts)) for pipeline in self.pipelines
         )
         return self